Moving CodeProblem model to model.dart
