+++ lib/src/fitness.dart
@@ -10,7 +10,6 @@ import 'dart:math';
 
 import 'package:path/path.dart' as p;
 
-import 'code_problem.dart';
 import 'messages.dart' as messages;
 import 'model.dart';
 import 'pubspec.dart';
